image_segmentation/pytorch/requirements.txt
git+https://github.com/NVIDIA/dllogger
https://github.com/mlcommons/logging/archive/refs/tags/1.1.0-rc4.zip
nibabel==3.2.1
scipy==1.5.2
tensorflow-cpu==2.9.1